requirements.txt
langchain>=0.1.12
langchain-openai>=0.0.8
langchain-community>=0.0.24
langchain-core>=0.1.32
openai>=1.12.0
python-dotenv==1.0.0
chromadb>=0.4.24
tiktoken>=0.6.0
